ai-caption-agent/requirements.txt
llama-cpp-python==0.2.11
requests==2.31.0
python-dotenv==1.0.0